drop non-preferred unstakes
